refactor: change e2e tests so its easier to use different providers
This change refactors the existing e2e tests so that there is a new spec
that can be re-used with different template files to enable us to test
different provider combinations.

Additionally the tests have been changed to use an in-cluster Git server
(gitea) so that we can create new repos as part of the test and remove
the reliance on the samples repo.

Signed-off-by: Richard Case <richard.case@suse.com>