feat: add CAPA e2e test
Added a new test that will use the AWS provider. It will provision a EKS
based cluster using managed node group.

Also labels have been added to the tests so that the tests can be
filtered with the `--test-filter` flag when running them. This is to
allow short and full tests.

Signed-off-by: Richard Case <richard.case@suse.com>