refactor: change azure resource group
This changes the e2e tests that create Azure cluster to use a specific
resource group. This will enable us to identify resource screated by us
for costs and also if we want to run a a janitor.

It also changed the embedded capi test to use Docker instead of Azure so
that we can save the costs of an additional Azure test.

Signed-off-by: Richard Case <richard.case@suse.com>