fix: installation on eks
There is an issue installing Rancher Turtles on EKS using the chart as
the kubernetes version in EKS doesn't follow semver. For example, EKS
will return a version like `v1.26.12-eks-5e0fdde` and this means that we
then try and pull the `docker.io/rancher/kubectl:v1.26.12-eks-5e0fdde`
image when performing the cleanup tasks.

The function to strip the suffix from the kubernetes version has been
updated to handle EKS, AKS and GKE.

Signed-off-by: Richard Case <richard.case@suse.com>