Remove used providers in e2e
EKS cluster suffers from limited number of private IPs, which are
exausted by test suites running in parallel. It affects GitRepo
processing, where a pod created to update content of the repo in order
to apply manifests to the cluster. Since the pod is not able to be
scheduled, e2e tests are failing.

Signed-off-by: Danil-Grigorev <danil.grigorev@suse.com>